Refactor definition of extra Lua modules
Instead of adding a separate file for each extra Lua module, let's group
them all together under the same macro definition. This way, we don't
need to patch the CE repository when we need to add a new Lua function
to EE.

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring
