+++ src/box/CMakeLists.txt
@@ -15,24 +15,6 @@ lua_source(lua_sources lua/session.lua session_lua)
 if (ENABLE_FEEDBACK_DAEMON)
     lua_source(lua_sources lua/feedback_daemon.lua feedback_daemon_lua)
 endif()
-if(ENABLE_AUDIT_LOG)
-    lua_source(lua_sources ${AUDIT_LOG_LUA_SOURCE} audit_lua)
-endif()
-if(ENABLE_SPACE_UPGRADE)
-    lua_source(lua_sources ${SPACE_UPGRADE_LUA_SOURCE} space_upgrade_lua)
-endif()
-if(ENABLE_FLIGHT_RECORDER)
-    lua_source(lua_sources ${FLIGHT_RECORDER_LUA_SOURCE} flightrec_lua)
-endif()
-if(ENABLE_READ_VIEW)
-    lua_source(lua_sources ${READ_VIEW_LUA_SOURCE} read_view_lua)
-endif()
-if(ENABLE_SECURITY)
-    lua_source(lua_sources ${SECURITY_LUA_SOURCE} security_lua)
-endif()
-if(ENABLE_INTEGRITY)
-    lua_source(lua_sources ${INTEGRITY_LUA_SOURCE} integrity_lua)
-endif()
 lua_source(lua_sources lua/net_box.lua net_box_lua)
 lua_source(lua_sources lua/net_replicaset.lua net_replicaset_lua)
 lua_source(lua_sources lua/upgrade.lua upgrade_lua)
@@ -76,16 +58,12 @@ lua_source(lua_sources lua/config/utils/tabulate.lua       config_utils_tabulate
 lua_source(lua_sources lua/config/utils/textutils.lua      config_utils_textutils_lua)
 lua_source(lua_sources lua/config/utils/funcutils.lua      config_utils_funcutils_lua)
 lua_source(lua_sources lua/config/utils/network.lua        config_utils_network_lua)
-
-if (ENABLE_CONFIG_EXTRAS)
-    lua_multi_source(lua_sources ${CONFIG_EXTRAS_LUA_SOURCES})
-endif()
 # }}} config
 
 lua_source(lua_sources lua/connpool.lua connpool_lua)
 
-if (ENABLE_FAILOVER)
-    lua_multi_source(lua_sources ${FAILOVER_LUA_SOURCES})
+if (ENABLE_BOX_LUA_EXTRAS)
+    lua_multi_source(lua_sources ${EXTRA_BOX_LUA_SOURCES})
 endif()
 
 set(bin_sources)
