+++ src/trivia/config.h.cmake
@@ -283,6 +283,8 @@
 /* Cacheline size to calculate alignments */
 #define CACHELINE_SIZE 64
 
+#cmakedefine ENABLE_LUA_EXTRAS 1
+#cmakedefine ENABLE_BOX_LUA_EXTRAS 1
 #cmakedefine ENABLE_ERRINJ_EXTRAS 1
 #cmakedefine ENABLE_FLIGHT_RECORDER 1
 #cmakedefine ENABLE_TUPLE_COMPRESSION 1
